+++ src/Tests/WB.Core.SharedKernels.DataCollection.Tests/InterviewTests/when_reevaluating_whole_interview_and_questionnaire_has_question_with_validation_expression.cs
@@ -52,7 +52,7 @@ namespace WB.Core.SharedKernels.DataCollection.Tests.InterviewTests
                 .Returns(expressionProcessor.Object);
 
             interview = CreateInterview(questionnaireId: questionnaireId);
-            interview.AnswerTextQuestion(userId, conditionallyInvalidQuestionId, new int[0], DateTime.Now, "answer");
+            interview.AnswerTextQuestion(userId, conditionallyInvalidQuestionId, new decimal[0], DateTime.Now, "answer");
 
             eventContext = new EventContext();
         };
